+++ src/gc.h
@@ -18,7 +18,7 @@ typedef struct external_heap heap_t;
 /// \param unsafe_stack true if pointers on the stack are to be considered unsafe pointers
 /// \param gc_threshold the memory pressure at which gc should be triggered (1.0 = full memory)
 /// \return the new heap
-heap_t *h_init(size_t bytes, bool unsafe_stack, float gc_threshold);
+heap_t *h_init(unsigned int bytes, bool unsafe_stack, float gc_threshold);
 
 /// Delete a heap.
 ///
@@ -57,7 +57,7 @@ void *h_alloc_struct(heap_t *h, char *layout);
 /// \param h the heap
 /// \param bytes the size in bytes
 /// \return the newly allocated object
-void *h_alloc_data(heap_t *h, size_t bytes);
+void *h_alloc_data(heap_t *h, unsigned int bytes);
 
 /// Manually trigger garbage collection.
 ///
@@ -66,7 +66,7 @@ void *h_alloc_data(heap_t *h, size_t bytes);
 ///
 /// \param h the heap
 /// \return the number of bytes collected
-size_t h_gc(heap_t *h);
+unsigned int h_gc(heap_t *h);
 
 /// Manually trigger garbage collection with the ability to 
 /// override the setting for how stack pointers are treated. 
@@ -77,13 +77,13 @@ size_t h_gc(heap_t *h);
 /// \param h the heap
 /// \param unsafe_stack true if pointers on the stack are to be considered unsafe pointers
 /// \return the number of bytes collected
-size_t h_gc_dbg(heap_t *h, bool unsafe_stack);
+unsigned int h_gc_dbg(heap_t *h, bool unsafe_stack);
 
 /// Returns the available free memory. 
 ///
 /// \param h the heap
 /// \return the available free memory. 
-size_t h_avail(heap_t *h);
+unsigned int h_avail(heap_t *h);
 
 /// Returns the bytes currently in use by user structures. This
 /// should not include the collector's own meta data. Notably,
@@ -92,6 +92,6 @@ size_t h_avail(heap_t *h);
 /// 
 /// \param h the heap
 /// \return the bytes currently in use by user structures. 
-size_t h_used(heap_t *h);
+unsigned int h_used(heap_t *h);
 
 #endif
