+++ src/heap.h
@@ -6,12 +6,12 @@
 
 typedef struct internal_heap internal_heap_t;
 
-internal_heap_t *h_init_internal(size_t bytes, size_t page_size);
+internal_heap_t *h_init_internal(unsigned int bytes, unsigned int page_size);
 
 void h_delete_internal(internal_heap_t *h);
 
 void *h_alloc_struct_internal(internal_heap_t *h, char *layout);
 
-void *h_alloc_data_internal(internal_heap_t *h, size_t bytes);
+void *h_alloc_data_internal(internal_heap_t *h, unsigned int bytes);
 
 #endif
